refactor: action types
