--- src/redux/actions/actionTypes.ts
@@ -1,17 +1,0 @@
-export const LOAD_GAME: string = "LOAD_GAME";
-export const GET_PIECES_STRENGTH: string = "GET_PIECES_STRENGTH";
-export const SELECT_PIECE: string = "SELECT_PIECE";
-export const CALCULATE_POSSIBLE_MOVES: string = "CALCULATE_POSSIBLE_MOVES";
-export const PIECE_MOVE: string = "PIECE_MOVE";
-export const USER_PIECE_MOVE: string = "USER_PIECE_MOVE";
-export const UPDATE_CURRENT_ARRANGEMENT_SEQ_NUM: string =
-  "UPDATE_CURRENT_ARRANGEMENT_SEQ_NUM";
-export const LOAD_BOARD_SUCCESS = "LOAD_BOARD_SUCCESS";
-export const LOADING_BOARD = "LOADING_BOARD";
-export const LOADING_BOARDS = "LOADING_BOARDS";
-export const LOAD_PIECES_SUCCESS = "LOAD_PIECES_SUCCESS";
-export const LOADING_PIECES = "LOADING_PIECES";
-export const LOAD_GAME_TYPE_SUCCESS = "LOAD_GAME_TYPE_SUCCESS";
-export const LOADING_GAME_TYPE = "LOADING_GAME_TYPE";
-export const SUBMIT_NEW_BOARD = "SUBMIT_NEW_BOARD";
-export const START_NEW_GAME = "START_NEW_GAME";
