+++ src/redux/actions/piecesActions.ts
@@ -1,5 +1,5 @@
 import * as api from "../../api";
-import { LOADING_PIECES, LOAD_PIECES_SUCCESS } from "./actionTypes";
+import { LOADING_PIECES, LOAD_PIECES_SUCCESS } from "../reducers/piecesReducer";
 
 export function loadPiecesByCodes(codes): any {
   return (dispatch) => {
