+++ src/redux/actions/playGameActions.ts
@@ -1,5 +1,4 @@
-// import * as api from "../../api";
-import { START_NEW_GAME } from "./actionTypes";
+export const START_NEW_GAME = "START_NEW_GAME";
 
 export function startNewGame(values) {
   return { type: START_NEW_GAME, payload: values };
